% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_apply_rsn.R
\name{apply_prm_style}
\alias{apply_prm_style}
\title{A \href{https://vega.github.io/vega-lite/}{vega-lite} theme formatted in the Refugee Solidarity Network style}
\usage{
apply_prm_style(spec = NULL, base_family = "Open Sans")
}
\arguments{
\item{spec, }{a vega-lite spec (typically passed in via pipe)}

\item{base_family, }{base font family}
}
\description{
\code{vl_apply_rsn} provides a \href{https://vega.github.io/vega-lite/}{vega-lite} theme formatted according to the
Refugee Solidarity Network style guide.
}
