% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_apply_map.R
\name{gg_apply_map}
\alias{gg_apply_map}
\title{A \link{ggplot2} theme formatted in the Refugee Solidarity Network style}
\usage{
gg_apply_map(
  scale = "continuous",
  base_size = 8.5,
  base_family = "Open Sans",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{scale}{"continuous" creates a vertical legend to the right of the map. "discrete" creates a horizontal legend above the map.}

\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{gg_apply_map} provides a \link{ggplot2} theme formatted according to the
RSN style guide for maps.
}
