% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_apply_rsn.R
\name{gg_apply_rsn}
\alias{gg_apply_rsn}
\title{A \link{ggplot2} theme formatted in the Refugee Solidarity Network style}
\usage{
gg_apply_rsn(
  base_size = 8.5,
  base_family = "Open Sans",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{gg_apply_rsn} provides a \link{ggplot2} theme formatted according to the
Refugee Solidarity Network style guide.
}
