% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_set_defaults.R
\name{gg_set_defaults}
\alias{gg_set_defaults}
\title{RSN \link{ggplot2} themes}
\usage{
gg_set_defaults(
  style = "rsn",
  base_size = 8.5,
  base_family = "Open Sans",
  base_line_size = 0.5,
  base_rect_size = 0.5,
  scale = "continuous"
)
}
\arguments{
\item{style}{The default theme style for the R session. "rsn", "ppr" or "map".}

\item{base_size}{The base font size for the theme. All fonts are relative to
this value.}

\item{base_family}{The base font family for the theme.}

\item{base_line_size}{The base line size for the theme. All line sizes are
relative to this value.}

\item{base_rect_size}{The base rect size for the theme. All rect sizes are
relative to this value.}

\item{scale}{For \code{theme_map()}. Should the legend theme be continuous or discrete?}
}
\description{
\code{gg_set_defaults} provides a \link{ggplot2} theme formatted according to the
RSN style guide.
}
