% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_sans_import.R
\name{open_sans_import}
\alias{open_sans_import}
\title{Import and register Open Sans font}
\usage{
open_sans_import()
}
\description{
\code{open_sans_import()} tests to see if Open Sans is imported and registered. If
Open Sans, isn't imported and registered, then \code{open_sans_import()} imports and
registers Open Sans with
\href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Open Sans must be installed on your computer for \code{open sans_import()} to
work. To install, visit
\href{https://fonts.google.com/specimen/Open+Sans}{Google fonts} and click
"Download family". Unzip and open each of the .ttf files and click install.

Test to see if Open Sans is imported and registered with \code{open sans_test()}.
}
