% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_sans_test.R
\name{open_sans_test}
\alias{open_sans_test}
\title{Test for Open Sans import and registration}
\usage{
open_sans_test()
}
\description{
\code{open_sans_test} tests to see if Open Sans is imported and registered. Open Sans is
can be installed from
\href{https://fonts.google.com/specimen/Open+Sans}{Google fonts}.
}
\details{
Import and register Open Sans in R with open_sans_install().
}
