% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_rsn.R
\docType{data}
\name{palette_rsn}
\alias{palette_rsn}
\title{The Refugee Solidarity Network \link{ggplot2} theme}
\format{
A \code{list}.
}
\usage{
palette_rsn
}
\description{
Palette data for RSN Style
}
\details{
Data used by the palettes in the rsnplots package.
}
\keyword{datasets}
