% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{rsn_color_pal}
\alias{rsn_color_pal}
\title{The Refugee Solidarity Network \link{ggplot2} theme}
\usage{
rsn_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used in the RSN theme.
}
