% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Discrete color scale that aligns with the RSN style}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{...}{other arguments passed to \code{discrete_scale()}}
}
\description{
This function can only handle up to 8 categories/colors.
}
\details{
If you need more than 8 colors for exploratory purposes, use
ggplot2::scale_color_discrete().
}
